Service provider refactoring
